refactor(contract): move governance and key-event votes into their own modules
