+++ crates/contract/src/api/key_events.rs
@@ -1,0 +1,268 @@
+//! Key-event lifecycle: starting and voting through keygen and resharing
+//! instances, and cancelling or aborting them.
+
+use crate::crypto_shared::types::PublicKeyExtendedConversionError;
+use crate::errors::{Error, InvalidParameters};
+use crate::primitives::key_state::KeyEventId;
+use crate::{MpcContract, MpcContractExt};
+use near_mpc_contract_interface::method_names;
+use near_mpc_contract_interface::types::{self as dtos};
+use near_sdk::{Gas, NearToken, Promise, env, log, near};
+
+#[near]
+impl MpcContract {
+    /// Starts a new attempt to generate a key for the current domain.
+    /// This only succeeds if the signer is the leader (the participant with the lowest ID).
+    #[handle_result]
+    pub fn start_keygen_instance(&mut self, key_event_id: KeyEventId) -> Result<(), Error> {
+        log!("start_keygen_instance: signer={}", env::signer_account_id(),);
+
+        self.assert_caller_is_attested_participant_and_protocol_active();
+
+        self.protocol_state
+            .start_keygen_instance(key_event_id, self.config.key_event_timeout_blocks)
+    }
+
+    /// Casts a vote for `public_key` for the attempt identified by `key_event_id`.
+    ///
+    /// The effect of this method is either:
+    ///  - Returns error (which aborts with no changes), if there is no active key generation
+    ///    attempt (including if the attempt timed out), if the signer is not a participant, or if
+    ///    the key_event_id corresponds to a different domain, different epoch, or different attempt
+    ///    from the current key generation attempt.
+    ///  - Returns Ok(()), with one of the following changes:
+    ///    - A vote has been collected but we don't have enough votes yet.
+    ///    - This vote is for a public key that disagrees from an earlier voted public key, causing
+    ///      the attempt to abort; another call to `start` is then necessary.
+    ///    - Everyone has now voted for the same public key; the state transitions into generating a
+    ///      key for the next domain.
+    ///    - Same as the last case, except that all domains have a generated key now, and the state
+    ///      transitions into Running with the newly generated keys.
+    #[handle_result]
+    pub fn vote_pk(
+        &mut self,
+        key_event_id: KeyEventId,
+        public_key: dtos::PublicKey,
+    ) -> Result<(), Error> {
+        log!(
+            "vote_pk: signer={}, key_event_id={:?}, public_key={:?}",
+            env::signer_account_id(),
+            key_event_id,
+            public_key,
+        );
+
+        self.assert_caller_is_attested_participant_and_protocol_active();
+
+        let extended_key =
+            public_key
+                .try_into()
+                .map_err(|err: PublicKeyExtendedConversionError| {
+                    InvalidParameters::MalformedPayload {
+                        reason: err.to_string(),
+                    }
+                })?;
+
+        if let Some(new_state) = self.protocol_state.vote_pk(key_event_id, extended_key)? {
+            self.protocol_state = new_state;
+        }
+
+        Ok(())
+    }
+
+    /// Starts a new attempt to reshare the key for the current domain.
+    /// This only succeeds if the signer is the leader (the participant with the lowest ID).
+    #[handle_result]
+    pub fn start_reshare_instance(&mut self, key_event_id: KeyEventId) -> Result<(), Error> {
+        log!(
+            "start_reshare_instance: signer={}",
+            env::signer_account_id()
+        );
+
+        self.assert_caller_is_attested_participant_and_protocol_active();
+        self.protocol_state
+            .start_reshare_instance(key_event_id, self.config.key_event_timeout_blocks)
+    }
+
+    /// Casts a vote for the successful resharing of the attempt identified by `key_event_id`.
+    ///
+    /// The effect of this method is either:
+    ///  - Returns error (which aborts with no changes), if there is no active key resharing attempt
+    ///    (including if the attempt timed out), if the signer is not a participant, or if the
+    ///    key_event_id corresponds to a different domain, different epoch, or different attempt
+    ///    from the current key resharing attempt.
+    ///  - Returns Ok(()), with one of the following changes:
+    ///    - A vote has been collected but we don't have enough votes yet.
+    ///    - Everyone has now voted; the state transitions into resharing the key for the next
+    ///      domain.
+    ///    - Same as the last case, except that all domains' keys have been reshared now, and the
+    ///      state transitions into Running with the newly reshared keys.
+    #[handle_result]
+    pub fn vote_reshared(&mut self, key_event_id: KeyEventId) -> Result<(), Error> {
+        log!(
+            "vote_reshared: signer={}, resharing_id={:?}",
+            env::signer_account_id(),
+            key_event_id,
+        );
+
+        self.assert_caller_is_attested_participant_and_protocol_active();
+
+        if let Some(new_state) = self.protocol_state.vote_reshared(key_event_id)? {
+            // Resharing has concluded, transition to running state
+            self.protocol_state = new_state;
+            self.recompute_available_foreign_chains();
+
+            // Spawn a promise to clean up votes from non-participants.
+            // Note: MpcContract::vote_update uses filtering to ensure correctness even if this cleanup fails.
+            Promise::new(env::current_account_id())
+                .function_call(
+                    method_names::REMOVE_NON_PARTICIPANT_UPDATE_VOTES.to_string(),
+                    vec![],
+                    NearToken::from_near(0),
+                    Gas::from_tgas(self.config.remove_non_participant_update_votes_tera_gas),
+                )
+                .detach();
+            // Spawn a promise to drop votes cast by non-participants.
+            Promise::new(env::current_account_id())
+                .function_call(
+                    method_names::CLEAN_TEE_STATUS.to_string(),
+                    vec![],
+                    NearToken::from_near(0),
+                    Gas::from_tgas(self.config.clean_tee_status_tera_gas),
+                )
+                .detach();
+            // Spawn a bounded sweep over stored attestations to prune invalid / expired entries.
+            Promise::new(env::current_account_id())
+                .function_call(
+                    method_names::CLEAN_INVALID_ATTESTATIONS.to_string(),
+                    serde_json::to_vec(&serde_json::json!({
+                        "max_scan": RESHARE_CLEAN_INVALID_ATTESTATIONS_MAX_SCAN
+                    }))
+                    .unwrap(),
+                    NearToken::from_near(0),
+                    Gas::from_tgas(self.config.clean_invalid_attestations_tera_gas),
+                )
+                .detach();
+            // Spawn a promise to clean up orphaned node migrations for non-participants
+            Promise::new(env::current_account_id())
+                .function_call(
+                    method_names::CLEANUP_ORPHANED_NODE_MIGRATIONS.to_string(),
+                    vec![],
+                    NearToken::from_near(0),
+                    Gas::from_tgas(self.config.cleanup_orphaned_node_migrations_tera_gas),
+                )
+                .detach();
+            // Spawn a promise to clean up foreign chain data for non-participants
+            Promise::new(env::current_account_id())
+                .function_call(
+                    method_names::CLEAN_FOREIGN_CHAIN_DATA.to_string(),
+                    vec![],
+                    NearToken::from_near(0),
+                    Gas::from_tgas(self.config.clean_foreign_chain_data_tera_gas),
+                )
+                .detach();
+            // Spawn a promise to drop verifier-change votes cast by non-participants
+            Promise::new(env::current_account_id())
+                .function_call(
+                    method_names::REMOVE_NON_PARTICIPANT_TEE_VERIFIER_VOTES.to_string(),
+                    vec![],
+                    NearToken::from_near(0),
+                    Gas::from_tgas(
+                        self.config
+                            .remove_non_participant_tee_verifier_votes_tera_gas,
+                    ),
+                )
+                .detach();
+        }
+
+        Ok(())
+    }
+
+    /// Casts a vote to cancel the current key resharing. If a threshold number of unique
+    /// votes are collected to cancel the resharing, the contract state will revert back to the
+    /// previous running state.
+    ///
+    /// - This method is idempotent, meaning a single account can not make more than one vote.
+    /// - Only nodes from the previous running state are allowed to vote.
+    ///
+    /// Return value:
+    /// - [Ok] if the vote was successfully collected.
+    /// - [Err] if:
+    ///     - The signer is not a participant in the previous running state.
+    ///     - The contract is not in a resharing state.
+    #[handle_result]
+    pub fn vote_cancel_resharing(&mut self) -> Result<(), Error> {
+        Self::assert_caller_is_signer();
+        log!("vote_cancel_resharing: signer={}", env::signer_account_id());
+
+        if let Some(new_state) = self.protocol_state.vote_cancel_resharing()? {
+            self.protocol_state = new_state;
+        }
+
+        Ok(())
+    }
+
+    /// Casts a vote to cancel key generation. Any keys that have already been generated
+    /// are kept and we transition into Running state; remaining domains are permanently deleted.
+    /// Deleted domain IDs cannot be reused again in future calls to vote_add_domains.
+    ///
+    /// A next_domain_id that matches that in the state's domains struct must be passed in. This is
+    /// to prevent stale requests from accidentally cancelling a future key generation state.
+    #[handle_result]
+    pub fn vote_cancel_keygen(&mut self, next_domain_id: u64) -> Result<(), Error> {
+        Self::assert_caller_is_signer();
+        log!("vote_cancel_keygen: signer={}", env::signer_account_id());
+
+        if let Some(new_state) = self.protocol_state.vote_cancel_keygen(next_domain_id)? {
+            self.protocol_state = new_state;
+        }
+        Ok(())
+    }
+
+    /// Casts a vote to abort the current key event instance. If succesful, the contract aborts the
+    /// instance and a new instance with the next attempt_id can be started.
+    #[handle_result]
+    pub fn vote_abort_key_event_instance(&mut self, key_event_id: KeyEventId) -> Result<(), Error> {
+        log!(
+            "vote_abort_key_event_instance: signer={}",
+            env::signer_account_id()
+        );
+
+        self.assert_caller_is_attested_participant_and_protocol_active();
+
+        self.protocol_state
+            .vote_abort_key_event_instance(key_event_id)
+    }
+}
+
+/// Entries to scan in the post-reshare `clean_invalid_attestations` sweep. External callers
+/// may pick a different value; this only governs the automatic invocation.
+///
+/// Sized against [`crate::config::Config::clean_invalid_attestations_tera_gas`]: scanning an
+/// entry costs gas whether or not it is removed, so raising this without raising that
+/// overruns the budget.
+const RESHARE_CLEAN_INVALID_ATTESTATIONS_MAX_SCAN: u32 = 30;
+
+#[cfg(not(target_arch = "wasm32"))]
+#[cfg(test)]
+#[expect(non_snake_case)]
+mod tests {
+    use crate::api::test_utils::forwarded_participant_call_contract;
+
+    #[test]
+    #[should_panic(expected = "Caller must be the signer account")]
+    fn vote_cancel_resharing__should_panic_when_predecessor_differs_from_signer() {
+        let mut contract = forwarded_participant_call_contract();
+        contract
+            .vote_cancel_resharing()
+            .expect("expected panic when predecessor != signer");
+    }
+
+    #[test]
+    #[should_panic(expected = "Caller must be the signer account")]
+    fn vote_cancel_keygen__should_panic_when_predecessor_differs_from_signer() {
+        let mut contract = forwarded_participant_call_contract();
+        contract
+            .vote_cancel_keygen(0)
+            .expect("expected panic when predecessor != signer");
+    }
+}
