+++ crates/contract/src/lib.rs
@@ -52,7 +52,7 @@ use crate::{
     update::{ProposedUpdates, Update, UpdateId},
 };
 use config::Config;
-use crypto_shared::types::{PublicKeyExtended, PublicKeyExtendedConversionError};
+use crypto_shared::types::PublicKeyExtended;
 use errors::{InvalidParameters, InvalidState, RespondError, TeeError};
 use near_mpc_contract_interface::deposits::{
     MINIMUM_NODE_MANAGEMENT_DEPOSIT_YOCTONEAR, SIGN_DEPOSIT_YOCTONEAR,
@@ -64,7 +64,7 @@ use near_mpc_contract_interface::types::{
     VerifyForeignTransactionRequestArgs, VerifyForeignTransactionResponse,
 };
 
-use dtos::{DomainConfig, DomainPurpose};
+use dtos::DomainPurpose;
 use mpc_attestation::attestation::{Attestation, DstackAttestation};
 use mpc_primitives::hash::{LauncherDockerComposeHash, LauncherImageHash, TeeVerifierCodeHash};
 use near_sdk::{
@@ -74,7 +74,7 @@ use near_sdk::{
 use node_migrations::NodeMigrations;
 use primitives::{
     domain::max_reconstruction_threshold,
-    key_state::{AuthenticatedParticipantId, EpochId, KeyEventId, Keyset},
+    key_state::{AuthenticatedParticipantId, Keyset},
     participants::ParticipantInfo,
     signature::{SignatureRequest, YieldIndex},
     thresholds::{
@@ -99,14 +99,6 @@ use tee::{
 /// node key cannot invoke these methods.
 pub const MINIMUM_NODE_MANAGEMENT_DEPOSIT: NearToken =
     NearToken::from_yoctonear(MINIMUM_NODE_MANAGEMENT_DEPOSIT_YOCTONEAR);
-
-/// Entries to scan in the post-reshare `clean_invalid_attestations` sweep. External callers
-/// may pick a different value; this only governs the automatic invocation.
-///
-/// Sized against [`crate::config::Config::clean_invalid_attestations_tera_gas`]: scanning an
-/// entry costs gas whether or not it is removed, so raising this without raising that
-/// overruns the budget.
-const RESHARE_CLEAN_INVALID_ATTESTATIONS_MAX_SCAN: u32 = 30;
 
 impl Default for MpcContract {
     fn default() -> Self {
@@ -173,10 +165,6 @@ impl SupportedForeignChainsByNode {
 }
 
 impl MpcContract {
-    fn threshold(&self) -> Result<GovernanceThreshold, Error> {
-        self.protocol_state.threshold()
-    }
-
     fn add_verify_foreign_tx_request(
         &mut self,
         request: VerifyForeignTransactionRequest,
@@ -563,99 +551,6 @@ impl MpcContract {
             }))
     }
 
-    #[expect(rustdoc::private_intra_doc_links)]
-    /// Propose new parameters for the MPC network: participants, governance
-    /// threshold, and optional per-domain
-    /// [`ReconstructionThreshold`](near_mpc_contract_interface::types::ReconstructionThreshold) updates
-    /// (empty map keeps the current ones), applied on resharing completion.
-    /// If a threshold number of votes are reached on the exact same proposal, this will transition
-    /// the contract into the Resharing state.
-    ///
-    /// The epoch_id must be equal to 1 plus the current epoch ID (if Running) or prospective epoch
-    /// ID (if Resharing). Otherwise the vote is ignored. This is to prevent late transactions from
-    /// accidentally voting on outdated proposals.
-    ///
-    /// Like the other governance voting methods, this must be called directly from the
-    /// participant's own NEAR account: [`assert_caller_is_signer()`](MpcContract::assert_caller_is_signer) requires
-    /// `signer_account_id == predecessor_account_id`, so calls forwarded through another
-    /// contract are rejected.
-    #[handle_result]
-    pub fn vote_new_parameters(
-        &mut self,
-        prospective_epoch_id: EpochId,
-        proposal: dtos::ProposedGovernanceThresholdParameters,
-    ) -> Result<(), Error> {
-        Self::assert_caller_is_signer();
-        let proposal: ProposedGovernanceThresholdParameters = proposal.try_into_contract_type()?;
-        log!(
-            "vote_new_parameters: signer={}, proposal={:?}",
-            env::signer_account_id(),
-            proposal,
-        );
-
-        let tee_upgrade_deadline_duration =
-            Duration::from_secs(self.config.tee_upgrade_deadline_duration_seconds);
-
-        let validation_result = self.tee_state.reverify_and_cleanup_participants(
-            proposal.participants(),
-            tee_upgrade_deadline_duration,
-        );
-
-        let proposed_participants = proposal.participants();
-        match validation_result {
-            TeeValidationResult::Full => {
-                if let Some(new_state) = self
-                    .protocol_state
-                    .vote_new_parameters(prospective_epoch_id, &proposal)?
-                {
-                    self.protocol_state = new_state;
-                }
-                Ok(())
-            }
-            TeeValidationResult::Partial {
-                participants_with_valid_attestation,
-            } => {
-                let invalid_participants: Vec<_> = proposed_participants
-                    .participants()
-                    .iter()
-                    .filter(|(account_id, _, _)| {
-                        !participants_with_valid_attestation
-                            .is_participant_given_account_id(account_id)
-                    })
-                    .collect();
-
-                Err(InvalidParameters::InvalidTeeRemoteAttestation {
-                    reason: format!(
-                        "The following participants have invalid TEE status: {:?}",
-                        invalid_participants
-                    ),
-                }
-                .into())
-            }
-        }
-    }
-
-    /// Propose adding a new set of domains for the MPC network.
-    /// If a threshold number of votes are reached on the exact same proposal, this will transition
-    /// the contract into the Initializing state to generate keys for the new domains.
-    ///
-    /// The specified list of domains must have increasing and contiguous IDs, and the first ID
-    /// must be the same as the `next_domain_id` returned by state().
-    #[handle_result]
-    pub fn vote_add_domains(&mut self, domains: Vec<DomainConfig>) -> Result<(), Error> {
-        Self::assert_caller_is_signer();
-        log!(
-            "vote_add_domains: signer={}, domains={:?}",
-            env::signer_account_id(),
-            domains,
-        );
-
-        if let Some(new_state) = self.protocol_state.vote_add_domains(domains)? {
-            self.protocol_state = new_state;
-        }
-        Ok(())
-    }
-
     /// Registers the set of foreign chains the calling node supports.
     ///
     /// Must be called directly from the participant's own NEAR account
@@ -769,228 +664,6 @@ impl MpcContract {
             .into();
 
         self.register_foreign_chain_support(foreign_chain_support)
-    }
-
-    /// Starts a new attempt to generate a key for the current domain.
-    /// This only succeeds if the signer is the leader (the participant with the lowest ID).
-    #[handle_result]
-    pub fn start_keygen_instance(&mut self, key_event_id: KeyEventId) -> Result<(), Error> {
-        log!("start_keygen_instance: signer={}", env::signer_account_id(),);
-
-        self.assert_caller_is_attested_participant_and_protocol_active();
-
-        self.protocol_state
-            .start_keygen_instance(key_event_id, self.config.key_event_timeout_blocks)
-    }
-
-    /// Casts a vote for `public_key` for the attempt identified by `key_event_id`.
-    ///
-    /// The effect of this method is either:
-    ///  - Returns error (which aborts with no changes), if there is no active key generation
-    ///    attempt (including if the attempt timed out), if the signer is not a participant, or if
-    ///    the key_event_id corresponds to a different domain, different epoch, or different attempt
-    ///    from the current key generation attempt.
-    ///  - Returns Ok(()), with one of the following changes:
-    ///    - A vote has been collected but we don't have enough votes yet.
-    ///    - This vote is for a public key that disagrees from an earlier voted public key, causing
-    ///      the attempt to abort; another call to `start` is then necessary.
-    ///    - Everyone has now voted for the same public key; the state transitions into generating a
-    ///      key for the next domain.
-    ///    - Same as the last case, except that all domains have a generated key now, and the state
-    ///      transitions into Running with the newly generated keys.
-    #[handle_result]
-    pub fn vote_pk(
-        &mut self,
-        key_event_id: KeyEventId,
-        public_key: dtos::PublicKey,
-    ) -> Result<(), Error> {
-        log!(
-            "vote_pk: signer={}, key_event_id={:?}, public_key={:?}",
-            env::signer_account_id(),
-            key_event_id,
-            public_key,
-        );
-
-        self.assert_caller_is_attested_participant_and_protocol_active();
-
-        let extended_key =
-            public_key
-                .try_into()
-                .map_err(|err: PublicKeyExtendedConversionError| {
-                    InvalidParameters::MalformedPayload {
-                        reason: err.to_string(),
-                    }
-                })?;
-
-        if let Some(new_state) = self.protocol_state.vote_pk(key_event_id, extended_key)? {
-            self.protocol_state = new_state;
-        }
-
-        Ok(())
-    }
-
-    /// Starts a new attempt to reshare the key for the current domain.
-    /// This only succeeds if the signer is the leader (the participant with the lowest ID).
-    #[handle_result]
-    pub fn start_reshare_instance(&mut self, key_event_id: KeyEventId) -> Result<(), Error> {
-        log!(
-            "start_reshare_instance: signer={}",
-            env::signer_account_id()
-        );
-
-        self.assert_caller_is_attested_participant_and_protocol_active();
-        self.protocol_state
-            .start_reshare_instance(key_event_id, self.config.key_event_timeout_blocks)
-    }
-
-    /// Casts a vote for the successful resharing of the attempt identified by `key_event_id`.
-    ///
-    /// The effect of this method is either:
-    ///  - Returns error (which aborts with no changes), if there is no active key resharing attempt
-    ///    (including if the attempt timed out), if the signer is not a participant, or if the
-    ///    key_event_id corresponds to a different domain, different epoch, or different attempt
-    ///    from the current key resharing attempt.
-    ///  - Returns Ok(()), with one of the following changes:
-    ///    - A vote has been collected but we don't have enough votes yet.
-    ///    - Everyone has now voted; the state transitions into resharing the key for the next
-    ///      domain.
-    ///    - Same as the last case, except that all domains' keys have been reshared now, and the
-    ///      state transitions into Running with the newly reshared keys.
-    #[handle_result]
-    pub fn vote_reshared(&mut self, key_event_id: KeyEventId) -> Result<(), Error> {
-        log!(
-            "vote_reshared: signer={}, resharing_id={:?}",
-            env::signer_account_id(),
-            key_event_id,
-        );
-
-        self.assert_caller_is_attested_participant_and_protocol_active();
-
-        if let Some(new_state) = self.protocol_state.vote_reshared(key_event_id)? {
-            // Resharing has concluded, transition to running state
-            self.protocol_state = new_state;
-            self.recompute_available_foreign_chains();
-
-            // Spawn a promise to clean up votes from non-participants.
-            // Note: MpcContract::vote_update uses filtering to ensure correctness even if this cleanup fails.
-            Promise::new(env::current_account_id())
-                .function_call(
-                    method_names::REMOVE_NON_PARTICIPANT_UPDATE_VOTES.to_string(),
-                    vec![],
-                    NearToken::from_near(0),
-                    Gas::from_tgas(self.config.remove_non_participant_update_votes_tera_gas),
-                )
-                .detach();
-            // Spawn a promise to drop votes cast by non-participants.
-            Promise::new(env::current_account_id())
-                .function_call(
-                    method_names::CLEAN_TEE_STATUS.to_string(),
-                    vec![],
-                    NearToken::from_near(0),
-                    Gas::from_tgas(self.config.clean_tee_status_tera_gas),
-                )
-                .detach();
-            // Spawn a bounded sweep over stored attestations to prune invalid / expired entries.
-            Promise::new(env::current_account_id())
-                .function_call(
-                    method_names::CLEAN_INVALID_ATTESTATIONS.to_string(),
-                    serde_json::to_vec(&serde_json::json!({
-                        "max_scan": RESHARE_CLEAN_INVALID_ATTESTATIONS_MAX_SCAN
-                    }))
-                    .unwrap(),
-                    NearToken::from_near(0),
-                    Gas::from_tgas(self.config.clean_invalid_attestations_tera_gas),
-                )
-                .detach();
-            // Spawn a promise to clean up orphaned node migrations for non-participants
-            Promise::new(env::current_account_id())
-                .function_call(
-                    method_names::CLEANUP_ORPHANED_NODE_MIGRATIONS.to_string(),
-                    vec![],
-                    NearToken::from_near(0),
-                    Gas::from_tgas(self.config.cleanup_orphaned_node_migrations_tera_gas),
-                )
-                .detach();
-            // Spawn a promise to clean up foreign chain data for non-participants
-            Promise::new(env::current_account_id())
-                .function_call(
-                    method_names::CLEAN_FOREIGN_CHAIN_DATA.to_string(),
-                    vec![],
-                    NearToken::from_near(0),
-                    Gas::from_tgas(self.config.clean_foreign_chain_data_tera_gas),
-                )
-                .detach();
-            // Spawn a promise to drop verifier-change votes cast by non-participants
-            Promise::new(env::current_account_id())
-                .function_call(
-                    method_names::REMOVE_NON_PARTICIPANT_TEE_VERIFIER_VOTES.to_string(),
-                    vec![],
-                    NearToken::from_near(0),
-                    Gas::from_tgas(
-                        self.config
-                            .remove_non_participant_tee_verifier_votes_tera_gas,
-                    ),
-                )
-                .detach();
-        }
-
-        Ok(())
-    }
-
-    /// Casts a vote to cancel the current key resharing. If a threshold number of unique
-    /// votes are collected to cancel the resharing, the contract state will revert back to the
-    /// previous running state.
-    ///
-    /// - This method is idempotent, meaning a single account can not make more than one vote.
-    /// - Only nodes from the previous running state are allowed to vote.
-    ///
-    /// Return value:
-    /// - [Ok] if the vote was successfully collected.
-    /// - [Err] if:
-    ///     - The signer is not a participant in the previous running state.
-    ///     - The contract is not in a resharing state.
-    #[handle_result]
-    pub fn vote_cancel_resharing(&mut self) -> Result<(), Error> {
-        Self::assert_caller_is_signer();
-        log!("vote_cancel_resharing: signer={}", env::signer_account_id());
-
-        if let Some(new_state) = self.protocol_state.vote_cancel_resharing()? {
-            self.protocol_state = new_state;
-        }
-
-        Ok(())
-    }
-
-    /// Casts a vote to cancel key generation. Any keys that have already been generated
-    /// are kept and we transition into Running state; remaining domains are permanently deleted.
-    /// Deleted domain IDs cannot be reused again in future calls to vote_add_domains.
-    ///
-    /// A next_domain_id that matches that in the state's domains struct must be passed in. This is
-    /// to prevent stale requests from accidentally cancelling a future key generation state.
-    #[handle_result]
-    pub fn vote_cancel_keygen(&mut self, next_domain_id: u64) -> Result<(), Error> {
-        Self::assert_caller_is_signer();
-        log!("vote_cancel_keygen: signer={}", env::signer_account_id());
-
-        if let Some(new_state) = self.protocol_state.vote_cancel_keygen(next_domain_id)? {
-            self.protocol_state = new_state;
-        }
-        Ok(())
-    }
-
-    /// Casts a vote to abort the current key event instance. If succesful, the contract aborts the
-    /// instance and a new instance with the next attempt_id can be started.
-    #[handle_result]
-    pub fn vote_abort_key_event_instance(&mut self, key_event_id: KeyEventId) -> Result<(), Error> {
-        log!(
-            "vote_abort_key_event_instance: signer={}",
-            env::signer_account_id()
-        );
-
-        self.assert_caller_is_attested_participant_and_protocol_active();
-
-        self.protocol_state
-            .vote_abort_key_event_instance(key_event_id)
     }
 
     /// Propose update to either code or config, but not both of them at the same time.
@@ -2198,9 +1871,9 @@ mod tests {
 
     use super::*;
     use crate::api::test_utils::*;
-    use crate::errors::DomainError;
-    use crate::errors::{InvalidCandidateSet, InvalidThreshold, NodeMigrationError};
+    use crate::errors::NodeMigrationError;
     use crate::primitives::domain::AddDomainsVotes;
+    use crate::primitives::key_state::EpochId;
     use crate::state::running::RunningContractState;
 
     use crate::primitives::participants::{ParticipantId, ParticipantInfo, Participants};
@@ -2831,362 +2504,6 @@ mod tests {
             contract.tee_verifier_votes(),
             BTreeMap::from([bucket(&candidate_b, &auth_b)]),
         );
-    }
-
-    /// Sets up the voting context and calls [`VersionedMpcContract::vote_new_parameters`] with the
-    /// given parameters.
-    fn setup_voting_context_and_vote(
-        contract: &mut MpcContract,
-        first_participant_id: &AccountId,
-        participants: Participants,
-        threshold: GovernanceThreshold,
-    ) -> Result<(), Error> {
-        let voting_context = VMContextBuilder::new()
-            .signer_account_id(first_participant_id.clone())
-            .predecessor_account_id(first_participant_id.clone())
-            .attached_deposit(NearToken::from_near(0))
-            .build();
-        testing_env!(voting_context);
-
-        let proposal = ProposedGovernanceThresholdParameters::new(
-            GovernanceThresholdParameters::new(participants, threshold).unwrap(),
-            BTreeMap::new(),
-        );
-        contract.vote_new_parameters(EpochId::new(1), (&proposal).into_dto_type())
-    }
-
-    /// Test that [`VersionedMpcContract::vote_new_parameters`] succeeds when all participants have
-    /// default TEE status ([`TeeQuoteStatus::None`]). This tests the basic scenario where no
-    /// participants have submitted attestation information, and all have the default TEE status
-    /// of [`TeeQuoteStatus::None`], which is considered acceptable.
-    #[test]
-    fn test_vote_new_parameters_succeeds_with_default_tee_status() {
-        let (mut contract, participants, first_participant_id) = setup_tee_test_contract(3, 2);
-        let threshold = GovernanceThreshold::new(2);
-
-        // No attestations submitted - all participants have default TEE status None
-        let result = setup_voting_context_and_vote(
-            &mut contract,
-            &first_participant_id,
-            participants,
-            threshold,
-        );
-        assert!(
-            result.is_ok(),
-            "Should succeed when all participants have default TEE status None"
-        );
-    }
-
-    /// Test that [`MpcContract::vote_new_parameters`] succeeds when all participants
-    /// submit valid TEE attestations. This tests the scenario where all participants successfully
-    /// submit valid attestations through [`MpcContract::submit_participant_info`],
-    /// resulting in [`TeeQuoteStatus::Valid`] TEE status for all participants.
-    #[test]
-    fn test_vote_new_parameters_succeeds_when_all_participants_have_valid_tee() {
-        let (mut contract, participants, first_participant_id) = setup_tee_test_contract(3, 2);
-        let threshold = GovernanceThreshold::new(2);
-
-        // Submit valid attestations for all participants
-        submit_valid_attestations(&mut contract, &participants, &[0, 1, 2]);
-
-        // This should succeed because all participants now have valid TEE status
-        let result = setup_voting_context_and_vote(
-            &mut contract,
-            &first_participant_id,
-            participants,
-            threshold,
-        );
-        assert!(
-            result.is_ok(),
-            "Should succeed when all participants have valid TEE status"
-        );
-    }
-
-    /// Test that attempts to submit invalid attestations are rejected by
-    /// [`MpcContract::submit_participant_info`]. This test demonstrates that
-    /// participants cannot have Invalid TEE status because the contract proactively rejects
-    /// invalid attestations at submission time. The 4th participant tries to submit an invalid
-    /// attestation but is rejected, leaving them with [`TeeQuoteStatus::Invalid`] status, which
-    /// combined with valid participants still allows successful voting.
-    #[test]
-    fn test_vote_new_parameters_succeeds_after_invalid_attestation_rejected() {
-        let (mut contract, participants, first_participant_id) = setup_tee_test_contract(4, 3);
-        let threshold = GovernanceThreshold::new(3);
-
-        // Submit valid attestations for first 3 participants
-        submit_valid_attestations(&mut contract, &participants, &[0, 1, 2]);
-
-        // Try to submit invalid attestation for the 4th participant
-        let participant_index = 3;
-        let result = submit_attestation(&mut contract, &participants, participant_index, false);
-        assert!(
-            result.is_err(),
-            "Invalid attestation should be rejected by submit_participant_info"
-        );
-
-        if let Err(error) = result {
-            let error_string = error.to_string();
-            assert!(
-                error_string.contains("failed verification"),
-                "Error should mention attestation verification failure, got: {}",
-                error_string
-            );
-        }
-
-        // This should succeed because:
-        // - 3 participants have Valid TEE status (from successful attestations)
-        // - 1 participant has None TEE status (invalid attestation was rejected)
-        // - Both Valid and None are allowed by the TEE validation
-        let result = setup_voting_context_and_vote(
-            &mut contract,
-            &first_participant_id,
-            participants,
-            threshold,
-        );
-        assert!(
-            result.is_ok(),
-            "Should succeed when participants have Valid or None TEE status (invalid attestations rejected)"
-        );
-    }
-
-    /// Builds a Running contract with `num_participants` participants, signing
-    /// threshold `threshold`, and a single CaitSith [`Sign`] domain whose
-    /// reconstruction threshold is `reconstruction_threshold`.
-    fn setup_running_contract_with_domain(
-        num_participants: usize,
-        threshold: u64,
-        reconstruction_threshold: u64,
-    ) -> (MpcContract, Participants, AccountId, DomainId) {
-        let participants = gen_participants(num_participants);
-        let first_participant_id = participants.participants()[0].0.clone();
-        testing_env!(
-            VMContextBuilder::new()
-                .signer_account_id(first_participant_id.clone())
-                .predecessor_account_id(first_participant_id.clone())
-                .attached_deposit(NearToken::from_near(1))
-                .build()
-        );
-
-        let parameters = GovernanceThresholdParameters::new(
-            participants.clone(),
-            GovernanceThreshold::new(threshold),
-        )
-        .unwrap();
-        let domain_id = DomainId::default();
-        let domains = vec![DomainConfig {
-            id: domain_id,
-            protocol: Protocol::CaitSith,
-            reconstruction_threshold: ReconstructionThreshold::new(reconstruction_threshold),
-            purpose: DomainPurpose::Sign,
-        }];
-        let (pk, _) = make_public_key_for_curve(Curve::Secp256k1, &mut OsRng);
-        let keyset = Keyset::new(
-            EpochId::new(0),
-            vec![KeyForDomain {
-                domain_id,
-                key: pk.try_into().unwrap(),
-                attempt: AttemptId::new(),
-            }],
-        );
-        let contract =
-            MpcContract::init_running(domains, 1, keyset, (&parameters).into_dto_type(), None)
-                .unwrap();
-        (contract, participants, first_participant_id, domain_id)
-    }
-
-    /// Installs a voting context for `signer` and casts `proposal`.
-    fn vote_params(
-        contract: &mut MpcContract,
-        signer: &AccountId,
-        proposal: &ProposedGovernanceThresholdParameters,
-    ) -> Result<(), Error> {
-        testing_env!(
-            VMContextBuilder::new()
-                .signer_account_id(signer.clone())
-                .predecessor_account_id(signer.clone())
-                .attached_deposit(NearToken::from_near(0))
-                .build()
-        );
-        contract.vote_new_parameters(EpochId::new(1), proposal.into_dto_type())
-    }
-
-    #[test]
-    fn vote_new_parameters__should_reject_when_per_domain_threshold_exceeds_participants() {
-        // Given: a Running contract with 3 participants and one domain.
-        let (mut contract, participants, signer, domain_id) =
-            setup_running_contract_with_domain(3, 2, 2);
-        // ...and a proposal raising that domain's reconstruction threshold to 4.
-        let mut per_domain = BTreeMap::new();
-        per_domain.insert(domain_id, ReconstructionThreshold::new(4));
-        let proposal = ProposedGovernanceThresholdParameters::new(
-            GovernanceThresholdParameters::new(participants, GovernanceThreshold::new(2)).unwrap(),
-            per_domain,
-        );
-
-        // When
-        let result = vote_params(&mut contract, &signer, &proposal);
-
-        // Then: 4 > 3 participants, so the guard rejects it.
-        assert_matches!(
-            result.unwrap_err(),
-            Error::DomainError(DomainError::ReconstructionThresholdExceedsParticipants {
-                reconstruction_threshold: 4,
-                participants: 3,
-            })
-        );
-    }
-
-    #[test]
-    fn vote_new_parameters__should_reject_when_shrinking_below_governance_threshold() {
-        // Given: a Running contract with 4 participants and a GovernanceThreshold of 3.
-        let (mut contract, participants, signer, _domain_id) =
-            setup_running_contract_with_domain(4, 3, 3);
-        // ...and a proposal that shrinks the participant set to 2 without touching
-        // the per-domain thresholds.
-        let proposal = ProposedGovernanceThresholdParameters::new(
-            GovernanceThresholdParameters::new(
-                participants.subset(0..2),
-                GovernanceThreshold::new(2),
-            )
-            .unwrap(),
-            BTreeMap::new(),
-        );
-
-        // When
-        let result = vote_params(&mut contract, &signer, &proposal);
-
-        // Then: the candidate-set guard rejects the proposal first, because only 2
-        // old participants remain — fewer than the GovernanceThreshold of 3. (Under
-        // the GovernanceThreshold >= max(ReconstructionThreshold) invariant this guard
-        // always fires before any per-domain ReconstructionThreshold check could.)
-        assert_matches!(
-            result.unwrap_err(),
-            Error::InvalidCandidateSet(InvalidCandidateSet::InsufficientOldParticipants)
-        );
-    }
-
-    #[test]
-    fn vote_new_parameters__should_reject_when_signing_threshold_exceeds_participants() {
-        // Given: a Running contract with 3 participants and one domain.
-        let (mut contract, participants, signer, _domain_id) =
-            setup_running_contract_with_domain(3, 2, 2);
-        // ...and a proposal whose signing threshold (4) exceeds the participant set.
-        let proposal = ProposedGovernanceThresholdParameters::new(
-            GovernanceThresholdParameters::new_unvalidated(
-                participants,
-                GovernanceThreshold::new(4),
-            ),
-            BTreeMap::new(),
-        );
-
-        // When
-        let result = vote_params(&mut contract, &signer, &proposal);
-
-        // Then
-        assert_matches!(
-            result.unwrap_err(),
-            Error::InvalidThreshold(InvalidThreshold::MaxRequirementFailed { max: 3, found: 4 })
-        );
-    }
-
-    #[test]
-    fn vote_new_parameters__should_accept_per_domain_threshold_within_participant_count() {
-        // Given: a Running contract with 5 participants (GovernanceThreshold 4) and one domain.
-        let (mut contract, participants, signer, domain_id) =
-            setup_running_contract_with_domain(5, 4, 2);
-        // ...and a proposal raising the domain's reconstruction threshold to 4,
-        // which fits the 5 participants and does not exceed the GovernanceThreshold.
-        let mut per_domain = BTreeMap::new();
-        per_domain.insert(domain_id, ReconstructionThreshold::new(4));
-        let proposal = ProposedGovernanceThresholdParameters::new(
-            GovernanceThresholdParameters::new(participants, GovernanceThreshold::new(4)).unwrap(),
-            per_domain,
-        );
-
-        // When: a single participant votes (no transition yet).
-        let result = vote_params(&mut contract, &signer, &proposal);
-
-        // Then: the guard passes and the vote is recorded.
-        assert_matches!(result, Ok(()));
-    }
-
-    #[test]
-    fn vote_new_parameters__should_reject_governance_below_max_reconstruction() {
-        // Given: a Running contract with 5 participants, GovernanceThreshold 4, and a
-        // domain whose reconstruction threshold is 4.
-        let (mut contract, participants, signer, _domain_id) =
-            setup_running_contract_with_domain(5, 4, 4);
-        // ...and a proposal lowering the GovernanceThreshold to 3 (valid on its own)
-        // while the domain keeps its reconstruction threshold of 4.
-        let proposal = ProposedGovernanceThresholdParameters::new(
-            GovernanceThresholdParameters::new(participants, GovernanceThreshold::new(3)).unwrap(),
-            BTreeMap::new(),
-        );
-
-        // When
-        let result = vote_params(&mut contract, &signer, &proposal);
-
-        // Then: the GovernanceThreshold (3) would fall below the domain's
-        // reconstruction threshold (4), so the proposal is rejected.
-        assert_matches!(
-            result.unwrap_err(),
-            Error::InvalidThreshold(InvalidThreshold::BelowReconstructionThreshold {
-                reconstruction_threshold: 4,
-                governance_threshold: 3,
-            })
-        );
-    }
-
-    #[test]
-    #[should_panic(expected = "Caller must be the signer account")]
-    fn vote_new_parameters__should_panic_when_predecessor_differs_from_signer() {
-        // Given: a participant whose vote is forwarded through another contract,
-        // so signer_account_id (the participant) != predecessor_account_id (the forwarder).
-        let (mut contract, participants, first_participant_id) = setup_tee_test_contract(3, 2);
-        let threshold = GovernanceThreshold::new(2);
-        let proposal = ProposedGovernanceThresholdParameters::new(
-            GovernanceThresholdParameters::new(participants, threshold).unwrap(),
-            BTreeMap::new(),
-        );
-
-        let ctx = VMContextBuilder::new()
-            .signer_account_id(first_participant_id)
-            .predecessor_account_id("forwarder.near".parse().unwrap())
-            .attached_deposit(NearToken::from_near(0))
-            .build();
-        testing_env!(ctx);
-
-        // When / Then: the confused-deputy vote must be rejected before it is recorded.
-        contract
-            .vote_new_parameters(EpochId::new(1), (&proposal).into_dto_type())
-            .expect("expected panic when predecessor != signer");
-    }
-
-    #[test]
-    #[should_panic(expected = "Caller must be the signer account")]
-    fn vote_add_domains__should_panic_when_predecessor_differs_from_signer() {
-        let mut contract = forwarded_participant_call_contract();
-        contract
-            .vote_add_domains(vec![])
-            .expect("expected panic when predecessor != signer");
-    }
-
-    #[test]
-    #[should_panic(expected = "Caller must be the signer account")]
-    fn vote_cancel_resharing__should_panic_when_predecessor_differs_from_signer() {
-        let mut contract = forwarded_participant_call_contract();
-        contract
-            .vote_cancel_resharing()
-            .expect("expected panic when predecessor != signer");
-    }
-
-    #[test]
-    #[should_panic(expected = "Caller must be the signer account")]
-    fn vote_cancel_keygen__should_panic_when_predecessor_differs_from_signer() {
-        let mut contract = forwarded_participant_call_contract();
-        contract
-            .vote_cancel_keygen(0)
-            .expect("expected panic when predecessor != signer");
     }
 
     #[test]
